Refactor all existing models
