+++ shardingsphere-v521-jdk8/src/main/java/com/lingh/shardingspherev521jdk8/mapper/TOrderShardingSphereMapper.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@SuppressWarnings({"SqlResolve", "SqlNoDataSourceInspection"})
+@SuppressWarnings({"SqlResolve", "SqlNoDataSourceInspection", "SqlDialectInspection"})
 @Component
 public interface TOrderShardingSphereMapper {
