+++ shardingsphere-v530-jdk8-legacy/src/main/java/com/lingh/shardingspherev530jdk8legacy/mapper/TOrderShardingSphereMapper.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@SuppressWarnings({"SqlResolve", "SqlNoDataSourceInspection"})
+@SuppressWarnings({"SqlResolve", "SqlNoDataSourceInspection", "SqlDialectInspection"})
 @Component
 public interface TOrderShardingSphereMapper {
