--- shardingsphere-v530-jdk8-modern/src/main/java/com/lingh/shardingspherev530jdk8modern/ShardingsphereV530Jdk8ModernApplication.java
@@ -1,15 +1,0 @@
-package com.lingh.shardingspherev530jdk8modern;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@MapperScan("com.lingh.shardingspherev530jdk8modern.mapper")
-public class ShardingsphereV530Jdk8ModernApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ShardingsphereV530Jdk8ModernApplication.class, args);
-	}
-
-}
