--- shardingsphere-v530-jdk8-modern/src/main/java/com/lingh/shardingspherev530jdk8modern/mapper/TOrderShardingSphereMapper.java
@@ -1,15 +1,0 @@
-package com.lingh.shardingspherev530jdk8modern.mapper;
-
-import com.lingh.shardingspherev530jdk8modern.pojo.TOrderShardingSpherePO;
-import org.apache.ibatis.annotations.Select;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@SuppressWarnings({"SqlResolve", "SqlNoDataSourceInspection"})
-@Component
-public interface TOrderShardingSphereMapper {
-
-    @Select("select * from t_order_sharding_sphere")
-    List<TOrderShardingSpherePO> findAll();
-}
