--- shardingsphere-v530-jdk8-modern/src/main/java/com/lingh/shardingspherev530jdk8modern/pojo/TOrderShardingSpherePO.java
@@ -1,12 +1,0 @@
-package com.lingh.shardingspherev530jdk8modern.pojo;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class TOrderShardingSpherePO {
-    private Long id;
-    private LocalDateTime createTime;
-    private String comment;
-}
