--- shardingsphere-v530-jdk8-modern/src/main/resources/application.yml
@@ -1,7 +1,0 @@
-mybatis:
-  configuration:
-    map-underscore-to-camel-case: true
-spring:
-  datasource:
-    driver-class-name: org.apache.shardingsphere.driver.ShardingSphereDriver
-    url: jdbc:shardingsphere:classpath:config.yaml
