--- shardingsphere-v530-jdk8-modern/src/test/java/com/lingh/shardingspherev530jdk8modern/ShardingsphereV530Jdk8ModernApplicationTests.java
@@ -1,13 +1,0 @@
-package com.lingh.shardingspherev530jdk8modern;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ShardingsphereV530Jdk8ModernApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
